Add files via upload
Dynamical Systems (DS) are a promising approach for robot motion modeling, which provide a flexible means to robot learning and control. Accuracy, stability and learning speed are the three major ingredients to learn robot motions from demonstrations with DS. Some approaches yield stable dynamical systems but potentially result in a poor reproduction performance, and some approaches yield good reproduction per- formance but are quite complex and time-consuming processes. Our focus in this paper is to address the accuracy-stability-speed issues simultaneously. We present a learning method named Fast and Stable Modeling for Dynamic Systems (FSM-DS), which is based on the Extreme Learning Machine to efficiently and accurately learn the parameters of the DS as will as to ensure the asymptotic stability at the target. The approach is validated through both 2-dimension tasks of learning handwriting motions and a set of robot experiments.